# libct/configs/validate: rootlessEUIDMount: speedup